moved to VM classes
