+++ cfme/cloud/instance/gce.py
@@ -1,4 +1,6 @@
 # -*- coding: utf-8 -*-
+from riggerlib import recursive_update
+
 from cfme.exceptions import OptionNotAvailable
 from cfme.utils import version, deferred_verpick
 from . import Instance
@@ -70,3 +72,23 @@ class GCEInstance(Instance):
             self.provider.mgmt.delete_vm(self.name)
         else:
             raise OptionNotAvailable(option + " is not a supported action")
+
+    @property
+    def vm_default_args(self):
+        """Represents dictionary used for Vm/Instance provision with GCE mandatory default args"""
+        inst_args = super(GCEInstance, self).vm_default_args
+        provisioning = self.provider.data['provisioning']
+        recursive_update(inst_args, {
+            'properties': {
+                'boot_disk_size': provisioning.get('boot_disk_size')}
+        })
+        return inst_args
+
+    @property
+    def vm_default_args_rest(self):
+        inst_args = super(GCEInstance, self).vm_default_args_rest
+        provisioning = self.provider.data['provisioning']
+        recursive_update(inst_args, {
+            'vm_fields': {
+                'boot_disk_size': provisioning['boot_disk_size'].replace(' ', '.')}})
+        return inst_args
